Fix timeouts
aiohttp timeouts, when specified as a single int, are for the total time, including any time that one might we waiting because of too many concurrent connections. When this library is being used in an environment like Home Assistant, where it might share a `ClientSession` with a bunch of other libraries, that can lead to spurious timeouts.

What we really wanted to was to time out quickly if the table doesn’t exist, so that’s a `sock_connect` timeout.